refactor: sequelize 관련 함수 common으로, 파일명 camel->pascal로 수정
