+++ libs/useApiNone.js
@@ -1,5 +1,5 @@
 const apiNoneCommands = require('./apiNoneCommands');
-const postMessageToSlack = require('../common/postMessageToSlack');
+const PostMessageToSlack = require('../common/PostMessageToSlack');
 
 // TODO 새로운 함수가 추가될 때마다 파일로 관리해서 기입하는 방식으로
 const partA = ['help', 'mail', 'github','suggest', 'lotto', 'menu', 'secret'];
